Renamed commands -> sdi and sdi->cli
This is so that imports work correctly and as expected
