+++ cli.py
@@ -50,4 +50,4 @@ def generator(func):
         yield from func(*args, **kwargs)
     return operator(update_wrapper(new_func, func))
 
-from commands import align, combine, subtract, snr_function, display, fitsio
+from sdi import align, combine, subtract, snr_function, display, fitsio
