+++ sdi/snr_function.py
